refactor: tidy up home
